New top level document concept and minor restructure for keywords
Top-level document element and accompanying modifications through the
system to make it work and fit together nicely.

Also some restructuring in code that works with keywords, since most
keywords relate to the document and it helps to make it explicit in
code which keywords do that and which keywords don't.

* lisp/org.el

** (new) org-keyword-regexp
Addition of regexp for keywords.  Let's try to use parameters and
constants instead of hardcoding these things all over the place...

** (removed) org-options-keywords.
This naming of this list is incongruent with the manual and the
concepts used throughout org-mode.  It also fits better in
org-element. So I'm removing this list in favour of new lists for
keywords in that file.

** (renamed, modified) org--setup-collect-keywords -> org-collect-keywords
Renamed and generalized org--setup-collect-keywords to make it work
for multiple purposes.  Is not limited to a fixed set of keywords any
longer.  New name: org-collect-keywords.

** (renamed, modified) org-make-options-regexp -> org-make-keyword-regexp
The EXTRA argument is never used so it is removed.  No need for
potential (or old?) functionatlity that adds complexity.  Can be used
together with org-collect-keywords to speed up processing.

** (renamed) org-set-regexps-and-options -> org-set-regexps-and-keywords
Renamed to more propertly match what it is doing.

* lisp/org-element.el
Addition of document as a concept and as a top-level node in the
content of org-data.  The content of a document element is everything
that previously was the content of org-data.

** org-element-*-keywords
Instead of org-options-keywords, a couple of lists are added to
org-element for keywords.  Multiple lists are used to indicate
different use-cases for particular sets of keywords.

*** (new) org-element-document-keywords
Keywords in this constant are those that apply for the whole document.

*** (new) org-element-export-keywords
List of keywords branded as export keywords.  These keywords are
supposed to be supported by all export backends.

*** (new) org-element-inline-keywords
List of keywords that provide content at a certain position in the
outline.

*** (new) org-element--get-document-keywords
Parses a document for keywords that are not dependent on their
position and should be regarded as applying to the whole document.

Return document and document export keywords based on reworked
function in org.el; org-collect-keywords.

** (modified) org-element-keyword-parser
Uses (new) org-keyword-regexp instead of hardcoding it's own regexp.

* lisp/ox.el
Fixing code so that it respects the new document element and uses
org-element functions instead of parsing org-element-contents itself.

* lisp/ob-core.el
Unused declaration for function, org-make-options-regexp, is removed.
It is removed since the function was renamed in lisp/org.el and found
here due to a dependency check.

* lisp/org-pcomplete.el
Cascading fixes due to cleanup in org and org-element.

* testing/lisp/test-org-element.el
Fix existing tests so they care for the document element.

* org.el