org, org-element, test-org, org-attach: Document property-drawers
* doc/org-manual.org

Describe the new possibility to define property-blocks on document
level, in addition to at headline level.

* etc/ORG-NEWS

Mention the change

* lisp/org.el

Code is modified to work with the concept of a node-level of 0 for
all text before the first headline.  Property drawers on
document-level, i.e node-level 0, work exactly the same as property
drawers on headlines.

Property drawers on document level can (only) be defined at the top of
a document.

** (Renamed) org-file-properties -> org-keyword-properties
Due to the fact that properties can be defined for the whole document
using property drawers this local variable needs a rename to make its
name less ambigous.

** (Modified) org-refresh-properties
Made to work also before first headline.

** (Modified) org-refresh-property
Made to work also before first headline.

** (Modified) org-refresh-category-properties
Made to work before first headline due to property drawers being
available there now!

** (Modified) org-get-property-block
Make it work also before first headline.

** (New) org-at-property-block-p
New function to validate if point is at the start of a property
block.

** (Modified) org-entry-properties
Make it work before first headline

** (Renamed) org-property-global-value -> org-property-global-or-keyword-value
Renamed to make its name match its functionality.

** (Modified) org-entry-get-with-inheritance
Make it work also before first headline.

** (Modified) org-entry-put
Make it work also before first headline.

** (Modified) org-insert-property-drawer
Make it work also before first headline.

** (New) org-back-to-heading-or-point-min
New function to make a document work as a level 0 node in the outline.

** (New) org-at-keyword-p
Predicate function to answer to if we're currently at a keyword line
or not.

** (New) org-up-heading-or-point-min
New function to make a document work as a level 0 node in the outline.

** (Modified) org-end-of-subtree
Make it work for the whole buffer, not just inside headlines. The end
of the buffer "subtree" is the end of the file.

* lisp/org-element.el

Add properties both from property drawer and property keywords to
document element. Drawer has precedence if properties are defined in
both.

Add post-blank property to document element  post-element is according
to ducumentation a property that all elements should have.

** (new) org-element--get-document-properties
Collect properties for whole documents from the property keywords and
from the document property drawer.

** (modified) org-element-document-parser
Add document properties to the document parser.

** (renamed) org-element--get-node-properties ->
   org-element--get-headline-properties
With the introduction of document property drawers while still
maintaining the concept of `node-property' for a property within a
property drawer, the function to get properties from headlines
property drawers need to be renamed to not be confusing.

(I suppose the concept of `node-property' could be removed in favour
of the more general `property' concept.  I couldn't calculate the
side-effects of doing that though.  And it would mean making changes in
far more files)

** (modified) org-element--current-element
Can now detect property-blocks before first headline according to it's
positional rules.

* lisp/org-attach.el

** (Modified) org-attach
Make it possible to call the attachment dispatcher (org-attach) also
before the first headline, since document property drawers make
attachments possible for the whole document now.

* lisp/org-capture.el

Modified only due to rename of function in org.el.

* testing/lisp/test-org.el

Add tests to make sure the specification of document-level
property-blocks can be maintained over time!

Additions to existing:
- /insert-property-drawer
- /set-property
- /delete-property
- /delete-property-globally

Added:
- /at-property-p
- /at-property-block-p
- /get-property-block
- /entry-get
- /refresh-properties

* testing/examples/property-inheritance.org

Switch from property-keywords to a property-drawer in the testfile.
Functionality should be the same, but now using a document drawer
instead of property-keywords.

Reason for switching is that I'd like us to slowly depricate
property-keywords.

* testing/lisp/test-org-element.el

Addition to existing test definitions:
- /node-property
- /property-drawer-parser

* contrib/lisp/ox-taskjuggler.el

A comment is modified only due to rename of function in org.el.